+++ include/sndboard.h
@@ -6,7 +6,7 @@
 bool toccata_init(struct autoconfig_info *aci);
 bool prelude_init(struct autoconfig_info *aci);
 bool prelude1200_init(struct autoconfig_info *aci);
-void update_sndboard_sound(double);
+void update_sndboard_sound(float);
 void sndboard_ext_volume(void);
 
 bool uaesndboard_init_z2(struct autoconfig_info *aci);
