+++ softfloat/softfloat_decimal.cpp
@@ -326,7 +326,7 @@ static int32_t getDecimalExponent(int32_t aExp, uint64_t aSig)
 		zSig0 &= ~(((int64_t)(zSig1<<1) == 0) & 1);
 	}
 	
-	zExp = zSign ? (0 - zSig0) : zSig0;
+	zExp = (int32_t)(zSign ? (0 - zSig0) : zSig0);
 
 	return zExp;
 }
